Enable wrapping on skill lists for mobile
This is only for ease of viewing purposes.
The downloaded resume (when implemented)
will have a standard sizing and margins without
wrapping. I figure it is better to have an
acceptable viewing experience on mobile, but
still keep the actual resume download up to
sizing standards. The other option was to
allow side scrolling on mobile and keep the
resume at full width, but that's not a good
viewing experience.